requirements-test.txt
django-nose==1.4.1
mock==1.3.0
nose==1.3.7
tox==2.1.1
